requirements.txt
alembic==1.7.7
black==22.3.0
click==8.1.2
Flask==2.1.1
Flask-JWT-Extended==4.4.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.20.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing-extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0
